% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{.galaxy_run_tool}
\alias{.galaxy_run_tool}
\title{Run a Galaxy tool programmatically}
\usage{
.galaxy_run_tool(
  tool_id,
  history_id,
  inputs,
  galaxy_url = "https://usegalaxy.eu"
)
}
\arguments{
\item{tool_id}{Character. Tool identifier to execute.}

\item{history_id}{Character. History ID where outputs will be stored.}

\item{inputs}{Named list describing tool inputs exactly as required
by the Galaxy tool. The list will be JSON-encoded automatically.}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
The job_id for the invocation.
}
\description{
Run a Galaxy tool programmatically
}
\details{
This sends a POST request to \code{/api/tools} with the required
payload. You can check the structure expected for \code{inputs} by
inspecting \code{galaxy_get_tool()} or by looking at Galaxy's
"Paste Request" feature in the web UI.
}
\examples{
\dontshow{if (galaxy_has_key()) withAutoprint(\{ # examplesIf}
history_id <- galaxy_initialize("add line")
tool <- galaxy_get_tool_id("Add line to file")
test_file <- tempfile(fileext = ".txt")
test_text <- "This is an example \ntest file."
writeLines(test_text,test_file)

file_id <- galaxy_upload_https(test_file, history_id)

inputs <- galaxy_get_tool(tool)

job_id <- galaxy_run_tool(tool_id = tool, history_id = history_id, inputs = list(
  text_input = "added text",
  infile = list(
    src = "hda",
    id  = file_id
  ),
  options = "header"   # optional, but explicit is good
))

result <- galaxy_poll_tool(job_id)

test_file_result <- tempfile(fileext = ".txt")
galaxy_download_result(list(output_ids = result$outputs$outfile$id), test_file_result)
readLines(test_file_result)
\dontshow{\}) # examplesIf}
}
