% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_get_workflow}
\alias{galaxy_get_workflow}
\title{Receive workflow metadata from the API}
\usage{
galaxy_get_workflow(workflow_id, galaxy_url = "https://usegalaxy.eu")
}
\arguments{
\item{workflow_id}{Character. Galaxy workflow ID.}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
a structured list with all metadata
}
\description{
Receive workflow metadata from the API
}
\examples{
\dontshow{if (nzchar(Sys.getenv("GALAXY_API_KEY"))) withAutoprint(\{ # examplesIf}
\dontrun{
galaxy_get_workflow("f2db41e1fa331b3e")
}
\dontshow{\}) # examplesIf}
}
