% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_has_key}
\alias{galaxy_has_key}
\title{Check whether a Galaxy API key is available}
\usage{
galaxy_has_key()
}
\value{
Logical. \code{TRUE} if an API key is available, otherwise \code{FALSE}.
}
\description{
Check whether the environment variable
\code{GALAXY_API_KEY} is set and non-empty.
}
\examples{
galaxy_has_key() # returns true if api key is set

}
