% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_initialize}
\alias{galaxy_initialize}
\title{Start a session with the GALAXY API}
\usage{
galaxy_initialize(name = "R API request", galaxy_url = "https://usegalaxy.eu")
}
\arguments{
\item{name}{Name of the history to create}

\item{galaxy_url}{Your Galaxy API key}
}
\value{
history_id The ID of the created history
}
\description{
Start a session with the GALAXY API
}
\examples{
\donotrun{
# set up your API Key in your .Renviron file first
galaxy_url <- "https://usegalaxy.eu"
api_key <- Sys.getenv("GALAXY_API_KEY")
history_id <- galaxy_initialize(api_key)
}
}
