% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_list_histories}
\alias{galaxy_list_histories}
\title{List Galaxy histories (name and history id)}
\usage{
galaxy_list_histories(galaxy_url = "https://usegalaxy.eu")
}
\arguments{
\item{galaxy_url}{Base URL of the Galaxy instance, e.g. "https://usegalaxy.eu"}
}
\value{
data.frame with columns: name, history_id
}
\description{
List Galaxy histories (name and history id)
}
