% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_list_invocations}
\alias{galaxy_list_invocations}
\title{List workflow invocations for a given workflow}
\usage{
galaxy_list_invocations(workflow_id, galaxy_url = "https://usegalaxy.eu")
}
\arguments{
\item{workflow_id}{The Galaxy workflow id to list invocations for}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
data.frame with columns: invocation_id, workflow_id, history_id, state, create_time, update_time
}
\description{
List workflow invocations for a given workflow
}
