% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_poll_tool,Galaxy-method}
\alias{galaxy_poll_tool,Galaxy-method}
\title{S4 method to poll the status of a tool invocation}
\usage{
\S4method{galaxy_poll_tool}{Galaxy}(
  x,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 3,
  timeout = 600,
  ...
)
}
\description{
S4 method to poll the status of a tool invocation
}
