% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_poll_workflow}
\alias{galaxy_poll_workflow}
\title{Poll a Galaxy workflow invocation until completion}
\usage{
galaxy_poll_workflow(
  invocation_id,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 30
)
}
\arguments{
\item{invocation_id}{The ID of the workflow invocation to poll}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}

\item{poll_interval}{Time in seconds between polling attempts in seconds}
}
\value{
A vector of HDA IDs corresponding to the output datasets of the workflow
}
\description{
Poll a Galaxy workflow invocation until completion
}
\examples{
\dontshow{if (galaxy_has_key()) withAutoprint(\{ # examplesIf}
# iris example
tmp_dir <- tempdir()
f_name <- "iris.csv"
f_path <- paste(tmp_dir, f_name, sep = "\\\\")
write.csv(datasets::iris, f_path, row.names = FALSE)

workflows <- galaxy_list_workflows(include_public = TRUE)
iris_workflow <- workflows[
  workflows$name == "Exploring Iris dataset with statistics and scatterplots",
][1,]

history_id <- galaxy_initialize("IRIS")
file_id <- galaxy_upload_https(f_path, history_id)
invocation_id <- galaxy_start_workflow(file_id, iris_workflow$id, history_id = history_id)
dataset_ids <- galaxy_poll_workflow(invocation_id)

result_files <- galaxy_get_file_info(dataset_ids$output_ids)
head(result_files)

\dontshow{\}) # examplesIf}
}
