% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_print_tool_inputs}
\alias{galaxy_print_tool_inputs}
\title{Print tool inputs}
\usage{
galaxy_print_tool_inputs(inputs)
}
\arguments{
\item{inputs}{A list of tool inputs as returned by \code{galaxy_get_tool}.}
}
\value{
A character vector with formatted input descriptions.
}
\description{
Print tool inputs
}
\examples{
\dontshow{if (galaxy_has_key()) withAutoprint(\{ # examplesIf}
tool_id <- galaxy_get_tool_id("FastQC")[1]
fastqc_tool <- galaxy_get_tool(tool_id)
print_tool_inputs(fastqc_tool)
\dontshow{\}) # examplesIf}
}
