% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_run_tool,Galaxy-method}
\alias{galaxy_run_tool,Galaxy-method}
\title{S4 Method for single tool invocation}
\usage{
\S4method{galaxy_run_tool}{Galaxy}(
  x,
  tool_id,
  inputs = NULL,
  galaxy_url = "https://usegalaxy.eu",
  ...
)
}
\description{
S4 Method for single tool invocation
}
