% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_upload}
\alias{galaxy_upload}
\title{Upload a file to Galaxy via FTP and register it in a history}
\usage{
galaxy_upload(
  input_file,
  history_id,
  galaxy_url = "https://usegalaxy.eu",
  galaxy_ftp = "ftp.usegalaxy.eu"
)
}
\arguments{
\item{input_file}{Path to the local file to upload}

\item{galaxy_url}{Base URL of the Galaxy instance}

\item{galaxy_ftp}{FTP server address of the Galaxy instance}
}
\value{
dataset_id The ID of the uploaded dataset in Galaxy
}
\description{
Upload a file to Galaxy via FTP and register it in a history
}
\examples{
\donotrun{
# set up your API Key, username and password in your .Renviron file first
galaxy_url <- "https://usegalaxy.eu"
galaxy_ftp <- "ftp.usegalaxy.eu"
input_file <- "path/to/your/file.txt"
dataset_id <- galaxy_upload(input_file, galaxy_url, galaxy_ftp)
print(dataset_id)
}
}
