% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_upload_ftp,Galaxy-method}
\alias{galaxy_upload_ftp,Galaxy-method}
\title{Galaxy upload via ftp S4 method}
\usage{
\S4method{galaxy_upload_ftp}{Galaxy}(
  x,
  input_file,
  galaxy_ftp = "ftp.usegalaxy.eu",
  galaxy_url = "https://usegalaxy.eu",
  ...
)
}
\description{
Galaxy upload via ftp S4 method
}
