% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_upload_https,Galaxy-method}
\alias{galaxy_upload_https,Galaxy-method}
\title{S4 Method for galaxy ftp upload}
\usage{
\S4method{galaxy_upload_https}{Galaxy}(
  x,
  input_file,
  wait = FALSE,
  wait_timeout = 600,
  galaxy_url = "https://usegalaxy.eu",
  file_type = "auto",
  dbkey = "?",
  ...
)
}
\description{
S4 Method for galaxy ftp upload
}
