% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_download_result,Galaxy-method}
\alias{galaxy_download_result,Galaxy-method}
\title{S4 download function for workflow or tool outputs}
\usage{
\S4method{galaxy_download_result}{Galaxy}(
  x,
  out_file = "result.laz",
  galaxy_url = "https://usegalaxy.eu",
  ...
)
}
\description{
S4 download function for workflow or tool outputs
}
