% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_download_result}
\alias{galaxy_download_result}
\title{Download final result dataset from Galaxy}
\usage{
galaxy_download_result(
  output_ids,
  out_file = "result.laz",
  galaxy_url = "https://usegalaxy.eu"
)
}
\arguments{
\item{output_ids}{Vector of HDA IDs from the workflow outputs the last one will be downloaded}

\item{out_file}{Path to save the downloaded file}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
The response object from the download request for debugging
}
\description{
Download final result dataset from Galaxy
}
