% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_download_result,Galaxy-method}
\alias{galaxy_download_result,Galaxy-method}
\title{S4 method to download files from a history}
\usage{
\S4method{galaxy_download_result}{Galaxy}(
  x,
  out_dir = ".",
  galaxy_url = "https://usegalaxy.eu",
  overwrite = FALSE,
  ...
)
}
\description{
S4 method to download files from a history
}
