% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_list_histories}
\alias{galaxy_list_histories}
\title{List Galaxy histories (name and history id)}
\usage{
galaxy_list_histories(galaxy_url = "https://usegalaxy.eu")
}
\arguments{
\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
data.frame with columns: history_name, history_id
}
\description{
List Galaxy histories (name and history id)
}
\examples{
\dontshow{if (galaxy_has_key()) withAutoprint(\{ # examplesIf}
histories <- galaxy_list_histories()
\dontshow{\}) # examplesIf}
}
