% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_poll_tool}
\alias{galaxy_poll_tool}
\alias{galaxy_poll_tool,character-method}
\alias{galaxy_poll_tool,Galaxy-method}
\title{Generic for galaxy_poll_tool}
\usage{
galaxy_poll_tool(
  x,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 3,
  timeout = 600
)

\S4method{galaxy_poll_tool}{character}(
  x,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 3,
  timeout = 600
)

\S4method{galaxy_poll_tool}{Galaxy}(x, poll_interval = 3, timeout = 600)
}
\arguments{
\item{x}{A job ID (\code{character}) or a \code{Galaxy} object.}

\item{galaxy_url}{Base URL of the Galaxy instance, used by the character
method.}

\item{poll_interval}{Seconds between status checks.}

\item{timeout}{Maximum time to wait in seconds.}
}
\value{
For the character method, the final job object; for the \code{Galaxy}
method, the modified \code{Galaxy} object.
}
\description{
Generic for galaxy_poll_tool

Wait for a Galaxy job to complete

S4 method to poll the status of a tool invocation
}
