% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_poll_workflow,Galaxy-method}
\alias{galaxy_poll_workflow,Galaxy-method}
\title{S4 object galaxy workflow polling function}
\usage{
\S4method{galaxy_poll_workflow}{Galaxy}(
  x,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 30,
  ...
)
}
\description{
S4 object galaxy workflow polling function
}
