% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_poll_workflow}
\alias{galaxy_poll_workflow}
\title{Poll a Galaxy workflow invocation until completion}
\usage{
galaxy_poll_workflow(
  invocation_id,
  galaxy_url = "https://usegalaxy.eu",
  poll_interval = 30
)
}
\arguments{
\item{invocation_id}{The ID of the workflow invocation to poll}

\item{galaxy_url}{Base URL of the Galaxy instance}

\item{poll_interval}{Time in seconds between polling attempts in seconds}
}
\value{
A vector of HDA IDs corresponding to the output datasets of the workflow
}
\description{
Poll a Galaxy workflow invocation until completion
}
