% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_start_workflow}
\alias{galaxy_start_workflow}
\title{Start a Galaxy workflow with a given dataset as input}
\usage{
galaxy_start_workflow(
  dataset_id,
  workflow_id,
  history_id = NA,
  galaxy_url = "https://usegalaxy.eu"
)
}
\arguments{
\item{dataset_id}{The ID of the input dataset in Galaxy}

\item{workflow_id}{The ID of the workflow to run}

\item{history_id}{The ID of the history where the workflow will run}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
invocation_id The ID of the started workflow invocation
}
\description{
Start a Galaxy workflow with a given dataset as input
}
