% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2026-01-21_s4_class_methods.R
\name{galaxy_start_workflow,Galaxy-method}
\alias{galaxy_start_workflow,Galaxy-method}
\title{S4 function to start a galaxy workflow}
\usage{
\S4method{galaxy_start_workflow}{Galaxy}(
  x,
  workflow_id,
  inputs = NULL,
  dataset_id = NULL,
  galaxy_url = "https://usegalaxy.eu",
  ...
)
}
\description{
S4 function to start a galaxy workflow
}
