% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_start_workflow}
\alias{galaxy_start_workflow}
\title{Start a Galaxy workflow with inputs and parameters}
\usage{
galaxy_start_workflow(
  dataset_id,
  workflow_id,
  history_id = NA,
  inputs = NULL,
  galaxy_url = "https://usegalaxy.eu"
)
}
\arguments{
\item{dataset_id}{Character. ID of the input dataset (HDA).
Used only if \code{inputs} is NULL.}

\item{workflow_id}{Character. Galaxy workflow ID.}

\item{history_id}{Character. History ID where the workflow will run.}

\item{inputs}{Named list. Optional workflow input mapping.
Keys must be workflow input step IDs (as characters).
Values must be lists describing datasets and/or parameters.}

\item{galaxy_url}{Character. Base URL of the Galaxy instance.}
}
\value{
Character. Workflow invocation ID.
}
\description{
Start a Galaxy workflow with inputs and parameters
}
