% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_upload_ftp}
\alias{galaxy_upload_ftp}
\title{FTP file upload to galaxy.}
\usage{
galaxy_upload_ftp(
  input_file,
  history_id,
  galaxy_ftp = "ftp.usegalaxy.eu",
  galaxy_url = "https://usegalaxy.eu"
)
}
\arguments{
\item{input_file}{Path to the local file to upload}

\item{history_id}{The ID of the Galaxy history where the dataset will be uploaded}

\item{galaxy_ftp}{FTP server address of the Galaxy instance}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}
}
\value{
dataset_id The ID of the uploaded dataset in Galaxy
}
\description{
Upload a file to Galaxy and register it in a history
}
\examples{
\dontshow{if (galaxy_has_key() & Sys.getenv("GALAXY_USERNAME") != "" & Sys.getenv("GALAXY_PASSWORD") != "") withAutoprint(\{ # examplesIf}
# set up your API Key, username and password in your .Renviron file first
galaxy_ftp <- "ftp.usegalaxy.eu"
input_file <- tempfile(fileext = ".txt")
test_text <- "This is an example \nfile."
writeLines(test_text,input_file)
history_id <- galaxy_initialize("test upload")
dataset_id <- galaxy_upload_ftp(input_file, history_id, galaxy_ftp)
print(dataset_id)
\dontshow{\}) # examplesIf}
}
