% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2025-10-27_JF_R_Galaxy_functions.R
\name{galaxy_upload_https}
\alias{galaxy_upload_https}
\title{Upload a dataset via HTTPS (direct POST) into Galaxy}
\usage{
galaxy_upload_https(
  input_file,
  history_id,
  wait = FALSE,
  wait_timeout = 600,
  galaxy_url = "https://usegalaxy.eu",
  file_type = "auto",
  dbkey = "?"
)
}
\arguments{
\item{input_file}{Character. Path to the local file to upload.}

\item{history_id}{Character. ID of the Galaxy history to receive the data set.}

\item{wait}{Logical. Whether to wait for Galaxy to finish processing}

\item{wait_timeout}{Integer. Time in seconds until wait times out with an error.}

\item{galaxy_url}{Character. Base URL of the Galaxy instance
(for example \code{"https://usegalaxy.eu"}).
If the environment variable \code{GALAXY_URL} is set, it takes precedence.}

\item{file_type}{Character. Galaxy datatype identifier
(for example \code{"auto"}, \code{"fastq"}, \code{"bam"}). Default: \code{"auto"}.}

\item{dbkey}{Character. Reference genome identifier (for example \code{"?"} or \code{"hg38"}). Default: \code{"?"}.}
}
\value{
A list describing the newly created dataset(s) as returned by Galaxy.
}
\description{
Upload a dataset via HTTPS (direct POST) into Galaxy
}
\details{
This function uses the built-in \code{upload1} tool and performs a
multipart form POST. Large files may still require FTP depending on
the Galaxy server's configuration limits.
}
\examples{
\dontshow{if (galaxy_has_key()) withAutoprint(\{ # examplesIf}
history_id <- galaxy_initialize("test upload")
test_file <- tempfile(fileext = ".txt")
test_text <- "This is an example \ntest file."
writeLines(test_text,test_file)

file_id <- galaxy_upload_https(test_file, history_id)
\dontshow{\}) # examplesIf}
}
